cleaning unnecessary imports

+++ src/app/api/logout/route.ts
@@ -1,17 +1,11 @@
-// import { singin } from "@/modules/auth/actions/auth-actions";
-// import { AuthService } from "@/modules/auth/services/auth-services";
-import { AuthService } from "@/app/module/services/auth-services";
+
+import { AuthService } from "@/app/module/auth/auth-services";
 import { createClient } from "@/utils/supabase/server";
-import { NextApiRequest, NextApiResponse } from "next";
 
-export async function DELETE(request: Request) {
-
-        const supabase = await createClient()
-        const session =  await AuthService.isSessionValid()
-        await supabase.auth.signOut()
-        if (session){
-            AuthService.deleteSession()
-        }
-    
-    return Response.json({logout:true})
+export async function DELETE() {
+    const supabase = await createClient()
+    const session = await AuthService.isSessionValid()
+    await supabase.auth.signOut()
+    if (session) AuthService.deleteSession()
+    return Response.json({ logout: true })
 }
